Fix definition of ALOHAOBJ with proper memory management
Before there was an error due to wrong management of the ALOHAOBJ
components memory wise. Now, both the 4-momentum and the wavefunction
arrays are instantiated beforehand as respectively real and complex,
then cast as fptype* both in the ALOHAOBJ constructor.
They are then accessed using the MemoryAccess headers in the usual way,
while being part of the ALOHAOBJ.
This approach is more consistent with the previous version of the code
and integrates the ALOHAOBJ class seamlessly.
There is no more the tentative of using the previous momenta memory and
reference it within the ALOHAOBJ 4-momentum, also because we need to do
some multiplication with the particle/antiparticle factors, and there
are also some +- signs to consider, and this should be done
independently from using the read momenta from the buffer inside the
HELAS routines.